Fixed bug in roundCoord where no coordinates were being rounded.
git-svn-id: http://ala-portal.googlecode.com/svn/trunk/biocache-store@2117 1ef1b718-b4f5-11dd-90a5-ed332e521da9

+++ src/main/scala/au/org/ala/biocache/caches.scala
@@ -393,7 +393,7 @@ object LocationDAO {
    */
   protected def roundCoord(x:String) : String = {
     try {
-      (((x * 10000).toInt).toFloat / 10000).toString
+      (((x.toFloat * 10000).toInt).toFloat / 10000).toString
     } catch {
       case e:NumberFormatException => x
     }
